+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - Changed the layout of `CbboMsg` to better match `BboMsg`
 - Renamed `Schema::Cbbo` to `Schema::Cmbp1`
 
+### Deprecations
+- Deprecated `Packing::Tar` and renamed it to `TarDeprecated`. This variant will be
+  removed in a future version when it is no longer supported by the API
+
 ## 0.22.0 - 2024-08-27
 
 ### Enhancements